pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "gpterm"
authors = [{name = "Ethan Ferguson", email = "sometimescode@gmail.com"}]
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]

dependencies = [
    "click",
    "openai>=1.4",
    "python-dotenv>=1",
    "readchar>=4",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "flake8",
    "black",
    "mypy",
    "isort",
    "pre-commit",
]

[project.scripts]
gpterm = "gpterm:main"
chat = "gpterm:main"
